move QuizService to services folder
