--- extensions/cli/src/quiz/index.ts
@@ -1,2 +1,0 @@
-export { QuizService, quizService } from "./QuizService.js";
-export * from "./types.js";
